+++ CanInterfaceImplementations/systec/STCanScan.h
@@ -55,7 +55,7 @@ public:
 
 	virtual int createBus(const std::string name ,const std::string parameters);
 	virtual int createBus(const std::string name, const std::string parameters, float factor );
-    virtual bool sendMessage(short cobID, unsigned char len, unsigned char *message, bool rtr = false, bool eff = false);
+    virtual bool sendMessage(uint32_t cobID, unsigned char len, unsigned char *message, bool rtr = false, bool eff = false);
 
     /*
 	 * Method that sends a remote request trough the can bus channel. If the method createBUS was not called before this, sendMessage will fail, as there is no
@@ -63,7 +63,7 @@ public:
 	 * @param cobID: Identifier that will be used for the request.
 	 * @return: Was the initialisation process successful?
 	 */
-	virtual bool sendRemoteRequest(short cobID);
+	virtual bool sendRemoteRequest(uint32_t cobID);
 	//Returns the instance of the CanStatistics object
 	virtual void getStatistics( CanStatistics & result );
 	// unified status
