+++ CanInterfaceImplementations/unitTestMockUpImplementation/MockCanAccess.h
@@ -47,8 +47,8 @@ public:
 	 */
 	virtual int createBus(const std::string name, const std::string parameters);
 	virtual int createBus(const std::string name, const std::string parameters, float factor );
-	virtual bool sendRemoteRequest(short cobID);
-    virtual bool sendMessage(short cobID, unsigned char len, unsigned char *message, bool rtr = false, bool eff = false);
+	virtual bool sendRemoteRequest(uint32_t cobID);
+    virtual bool sendMessage(uint32_t cobID, unsigned char len, unsigned char *message, bool rtr = false, bool eff = false);
 	virtual void getStatistics( CanModule::CanStatistics & result );
 	virtual bool initialiseLogging(LogItInstance* remoteInstance);
